requirements.txt
Flask==3.0.2
Flask-Cors==4.0.1
websocket-client==1.7.0
python-dotenv==1.0.1
requests==2.31.0
boto3==1.34.48
botocore==1.34.48